Implement ability to add new Breeds
There's lots of shortcuts on the frontend and controller pieces of
this. The form + the controller just treat it as a basic HTML form
upload, ending with a redirect on success. This isn't perfect but
is much simpler than handling file uploads in React.

Also relatively light on error handling logic.

Next step on upload features is to split the Uploader to support
S3 compatible uploads. The groundwork is already there to do this.