Initial Commit; seems bigger than it is
Generated a new Phoenix application for this challenge.
The main points of interest are the ImageUploader and the DogController.

Initially I wrote it so that it would just store images locally, but
when I went to get it released, I realized I didn't have the ability to
write to disk on Fly.io. So I went ahead and added ExAWS and set up
an S3 bucket, and all the configuration to allow that to work.
Despite it working splendidly locally, it would not work on Fly.io due
to some env var configuration issue.